moved header card to its own component
